Move test-only FeatureList init to features_test_util
